Moved Compile Seeds function to /utils for testing.
