+++ src/utils.js
@@ -32,3 +32,56 @@ export const getLongestStreak = (winlosses, matchString) => {
   }, "");
   return longestStreak.length;
 };
+
+export const compileWeeklySeeds = (owners, allMatchupsInYear) => {
+        const clonedOwners = [...owners];
+
+        allMatchupsInYear.forEach((matchupWeek, weekIndex) => {
+            const matchupMap = {};
+
+            // Group matchups by matchup_id
+            matchupWeek.forEach(team => {
+                const id = team.matchup_id;
+                if (!matchupMap[id]) matchupMap[id] = [];
+                matchupMap[id].push(team);
+            });
+
+            // Compare each matchup pair
+            Object.values(matchupMap).forEach(pair => {
+                if (pair.length === 2) {
+                    const [teamA, teamB] = pair;
+
+                    const ownerA = clonedOwners.find(o => o.rosterId === teamA.roster_id);
+                    const ownerB = clonedOwners.find(o => o.rosterId === teamB.roster_id);
+
+                    if (!ownerA || !ownerB) return;
+
+                    ownerA.pointsCount += teamA.points;
+                    ownerB.pointsCount += teamB.points;
+
+                    if (teamA.points > teamB.points) {
+                        ownerA.winCount += 1;
+                    } else if (teamB.points > teamA.points) {
+                        ownerB.winCount += 1;
+                    }
+                }
+            });
+
+            // Rank all owners by winCount, then pointsCount
+            const rankings = [...clonedOwners]
+                .sort((a, b) => {
+                    if (b.winCount === a.winCount) {
+                        return b.pointsCount - a.pointsCount;
+                    }
+                    return b.winCount - a.winCount;
+                });
+
+            // Store weekly seeds
+            rankings.forEach((owner, i) => {
+                owner.seeds[`Week ${weekIndex + 1} Seed`] = i + 1;
+            });
+        });
+
+        console.log("Compiled Owners with Weekly Seeds:", clonedOwners);
+        return clonedOwners;
+    };
